backend/requirements.txt
notebooklm-py
fastapi
uvicorn[standard]
httpx
python-dotenv
